OPCT-276: bump to v0.6.0-alpha.2 to collect kube-burner data
This change introduces kube-burner performance data collection.

kube-burner is a standard tool to collect performance and scale tests at
OpenShift, those profiles would be the basic one to evaluate the current
cluster state with small size.

It's not the goal to extract insights, we need first to collect data
points to a controlled environment before proceeding with any
evaluation/report.